==> _previous/clustering/albou558/can_2025/bourdieu/02-2_preparation_donnees_style.ipynb <==
# Préparation des variables pour les variables de style

library(tidyverse)
library(factoextra)

# Charger les données
DataPilot <- readRDS("../../_PrivateFolder_datagotchi_federal_2025/data/clustering/can2025/01_pilot_2025.rds")

# Charger le fichier des variables
source("01-1_variables.R")

# Préparation du jeu de données pour les variables de style
df_style <- DataPilot %>%
  select(all_of(variables_style))

# Boucle pour afficher la table pour chaque variable
for (v in variables_style) {
  if (v != "id") {
    cat("Table pour la variable:", v, "\n")
    print(table(df_style[[v]]))
    cat("\n\n")
  }
}

## Transformations

- app_swag

Les variables app_swag sont telles que certaines catégories sont très peu peuplées. Regrouper les styles pour en faire des styles plus généraux.

- animal

Regrouper certaines classes d'animal pour avoir des catégories plus générales


# Regroupement de app_swag
#-------------------------

df_style <- df_style %>%

  # Regrouper Formel et Chic sous Classique
  mutate(
    lifestyle_clothingStyleClassic = if_else(
      lifestyle_clothingStyleFormal == 1 | lifestyle_clothingStyleElegant == 1,
      1,
      lifestyle_clothingStyleClassic
    )
  ) %>%
  select(-lifestyle_clothingStyleFormal, -lifestyle_clothingStyleElegant) %>%

  # Regrouper HippBoheme, Punk, Rock sous Other
  mutate(
    lifestyle_clothingStyleOther = if_else(
      lifestyle_clothingStyleHippie == 1 | lifestyle_clothingStylePunk == 1 | lifestyle_clothingStyleRock == 1,
      1,
      lifestyle_clothingStyleOther
    )
  ) %>%
  select(-lifestyle_clothingStyleHippie, -lifestyle_clothingStylePunk, -lifestyle_clothingStyleRock)

# Regroupement de animal
#-----------------------

df_style <- df_style %>%
  mutate(lifestyle_ownPetOther = if_else(
   lifestyle_ownPetFarmAnimals == 1,
    1,
    lifestyle_ownPetOther
  )
  ) %>%
  select(-lifestyle_ownPetFarmAnimals)

variables_style_clust <- c(
  "id",
  "lifestyle_clothingStyleClassic",
  "lifestyle_clothingStyleCasual",
  "lifestyle_clothingStyleSport",
  "lifestyle_clothingStyleOther",
  "lifestyle_hasTattoos",
  "lifestyle_ownPetCat",
  "lifestyle_ownPetDog",
  "lifestyle_ownPetOther",
  "lifestyle_ownPetCatAndDog",
  "lifestyle_ownPetNone"
)

# Boucle pour afficher la table pour chaque variable
for (v in variables_style_clust) {
  if (v != "id") {
    cat("Table pour la variable:", v, "\n")
    print(table(df_style[[v]]))
    cat("\n\n")
  }
}

# Sauvegarder les données préparées
saveRDS(df_style, file = "../../_PrivateFolder_datagotchi_federal_2025/data/clustering/can2025/02_pilot_2025_style.rds")

---

# SUGGESTION : Activité de Clustering

Si souhaité, effectuer un exercice de clustering avec ce sous-groupe de variables pour mieux saisir les données et leur relation. 

Cet exercice est purement exploratoire et sert à voir si les données peuvent être utilisées pour faire du clustering.
