==> _previous/clustering/albou558/can_2025/bourdieu/02-7_preparation_donnees_finales.ipynb <==
# Assemblage des données

Les variables ont été préparées pour le clustering en sous-ensemble de variables.

Assembler toutes les tables de données contenant chaque sous-ensemble de variables.

library(dplyr)

# Assemblage des fichiers

df_act         <- readRDS(file = "../../_PrivateFolder_datagotchi_federal_2025/data/clustering/can2025/02_pilot_2025_act.rds")
df_style       <- readRDS(file = "../../_PrivateFolder_datagotchi_federal_2025/data/clustering/can2025/02_pilot_2025_style.rds")
df_sante       <- readRDS(file = "../../_PrivateFolder_datagotchi_federal_2025/data/clustering/can2025/02_pilot_2025_sante.rds")
df_mode_de_vie <- readRDS(file = "../../_PrivateFolder_datagotchi_federal_2025/data/clustering/can2025/02_pilot_2025_mode_de_vie.rds")
df_commerce    <- readRDS(file = "../../_PrivateFolder_datagotchi_federal_2025/data/clustering/can2025/02_pilot_2025_commerce.rds")
df_ses         <- readRDS(file = "../../_PrivateFolder_datagotchi_federal_2025/data/clustering/can2025/02_pilot_2025_ses.rds")

df_pilot_2025_clust <- df_act %>%
  left_join(df_style, by = "id") %>%
  left_join(df_sante, by = "id") %>%
  left_join(df_mode_de_vie, by = "id") %>%
  left_join(df_commerce, by = "id") %>%
  left_join(df_ses, by = "id")

saveRDS(df_pilot_2025_clust, file = "../../_PrivateFolder_datagotchi_federal_2025/data/clustering/can2025/02_pilot_2025.rds")
