==> _previous/clustering/albou558/can_2025/bourdieu/03-0_variables.R <==
#+++++++++++++++++++++++++++++++++++++++++++++++++++++++++
# 03 - Variables à utiliser pour le clustering
#+++++++++++++++++++++++++++++++++++++++++++++++++++++++++

library(dplyr)

# Déclaration des variables de clustering
variables_clust_bourdieu <- c("score_diff", "score_sum")
