==> _previous/clustering/albou558/can_2025/bourdieu/03-5_desc_geo.ipynb <==
library(sf)
library(ggplot2)
library(patchwork)
library(dplyr)

data_pilote2025<- readRDS(file ="../../_PrivateFolder_datagotchi_federal_2025/data/clustering/can2025/03_pilot_2025.rds")

names(data_pilote2025)

DataPilot <- readRDS("../../_SharedFolder_datagotchi_federal_2024/data/pilote/dataClean/datagotchi2025_canada_pilotClustering_20250305.rds")

names(DataPilot)
DataPilot$ses_postalCode

library(dplyr)

sf_rta <- cartessn::spatial_canada_2021_rta

sf_ridings <- cartessn::spatial_canada_2022_electoral_ridings

df_int <- cartessn::intersect_spatial_objects(
  sf_rta,
  "rta",
  sf_ridings,
  "id_riding"
) |> 
  arrange(rta) |> 
  select(rta, id_riding, prop_of_rta_area_covered_by_riding = prop_of_ref_area_covered_by_target)

# On récupère la carte des circonscriptions de 2022
map_data <- cartessn::spatial_canada_2022_electoral_ridings

# On repasse en projection lat/lon
map_data <- st_transform(map_data, crs = 4326)


p_main <- ggplot(map_data) +
  geom_sf(aes(fill = my_variable), color = NA) +
  # Choisissez la palette de couleurs qui vous convient :
  scale_fill_viridis_d(option = "C", na.value = "grey80") +
  theme_minimal() +
  labs(
    title = "Carte globale des circonscriptions canadiennes",
    fill = "Ma variable"  # Légende
  )


# Québec
quebec_data <- cartessn::crop_map(map_data, "quebec_city")
p_quebec <- ggplot(quebec_data) +
  geom_sf(aes(fill = my_variable), color = NA) +
  scale_fill_viridis_d(option = "C", na.value = "grey80") +
  theme_void() +
  labs(title = "Québec")

# Montréal (souvent "montreal" ou "montreal_region")
montreal_data <- cartessn::crop_map(map_data, "montreal")
p_montreal <- ggplot(montreal_data) +
  geom_sf(aes(fill = my_variable), color = NA) +
  scale_fill_viridis_d(option = "C", na.value = "grey80") +
  theme_void() +
  labs(title = "Montréal")

# Vancouver
vancouver_data <- cartessn::crop_map(map_data, "vancouver")
p_vancouver <- ggplot(vancouver_data) +
  geom_sf(aes(fill = my_variable), color = NA) +
  scale_fill_viridis_d(option = "C", na.value = "grey80") +
  theme_void() +
  labs(title = "Vancouver")


# Winnipeg
winnipeg_data <- cartessn::crop_map(map_data, "winnipeg")
p_winnipeg <- ggplot(winnipeg_data) +
  geom_sf(aes(fill = my_variable), color = NA) +
  scale_fill_viridis_d(option = "C", na.value = "grey80") +
  theme_void() +
  labs(title = "Winnipeg")

# Toronto (souvent "toronto" ou "gta" ou "toronto_gta")
gta_data <- cartessn::crop_map(map_data, "toronto")
p_gta <- ggplot(gta_data) +
  geom_sf(aes(fill = my_variable), color = NA) +
  scale_fill_viridis_d(option = "C", na.value = "grey80") +
  theme_void() +
  labs(title = "GTA / Toronto")


# On combine le tout.
# Ici, on fait une disposition verticale (ncol = 1) :
p_final <- p_main + 
  (p_quebec + p_montreal + p_vancouver) +
  (p_winnipeg + p_gta) +
  plot_layout(ncol = 1)

# On affiche le résultat
p_final
